QuranPage, FetchingDataError, directory DataModels
Отделил QuranPage struct и enum FetchingDataError от QuranData и поместил их в папку DataModels
