+++ QuranTracker/Model/DataModels/FetchingDataError.swift
@@ -1,0 +1,12 @@
+//
+//  FetchingDataError.swift
+//  QuranTracker
+//
+//  Created by Gamıd Khalıdov on 12.11.2024.
+//
+
+import Foundation
+
+enum FetchingDataError: Error {
+    case invalidPageNumber, invalidURL, errorWhileDecodingData, InvalidResponse, InvalidData
+}
