+++ QuranTracker/Model/QuranData.swift
@@ -34,34 +34,3 @@ class QuranData {
     
 
 }
-
-struct DataWrapper: Codable {
-    let data: QuranPage
-}
-
-
-// MARK: - QuranPage
-struct QuranPage: Codable {
-    let number: Int
-    let surahs: [String: Surah]
-    
-    static let mock = QuranPage(
-        number: 2,
-        surahs: ["2" : .mock]
-    )
-}
-
-// MARK: - Surah
-struct Surah: Codable {
-    let number: Int
-    let englishName: String
-    
-    static let mock = Surah(
-        number: 2,
-        englishName: "Al-Baqara"
-    )
-}
-
-enum FetchingDataError: Error {
-    case invalidPageNumber, invalidURL, errorWhileDecodingData, InvalidResponse, InvalidData
-}
